replaced fill_arena with load_memory wich does exactly the same, it's basically just a rename and a few little optimisations - added a check in parser to make sure that the user doesnt add a champ with a higher number than the actual amount of champs - added a sort_champ_tab to .. sort the champ_tab wich makes the memory loading easier

+++ includes/typedef.h
@@ -6,7 +6,7 @@
 /*   By: paullaurent <paullaurent@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/07 10:30:02 by tlandema          #+#    #+#             */
-/*   Updated: 2019/11/28 14:03:54 by tlandema         ###   ########.fr       */
+/*   Updated: 2019/12/02 17:40:52 by brichard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -142,7 +142,7 @@ typedef struct		s_parser
 {
 	t_vm			env;
 	enum e_state	state;
-	int32_t			chp_num;
+	uint8_t			chp_num;
 	uint8_t			cur_chp_index;
 }					t_parser;
